+++ dot_sh/session/pyenv.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env sh
+
+if command -v pyenv >/dev/null 2>&1 ; then
+  eval "$(pyenv init -)"
+fi